renamed comment component file and relocated to interactions directory
